refactor: Update layout and not-found page (#25)
* style(layout): Standardize layout padding

Refactor padding in the main application layout to use a centralized CSS
variable.

The `--header-height` variable is added to `globals.css` and applied
to `layout-wrapper.tsx` for consistent spacing and
improved maintainability.

* refactor(not-found): Update 404 content and styles

The content and styling of the existing 404 page have been updated to
provide a more user-friendly experience for invalid routes.
The changes also include wrapping the page in a `<PageContainer>`
component and modifying the text and elements for improved guidance.

* style(tailwind): Reorder classnames

Reorder Tailwind CSS classnames in `layout-wrapper.tsx` to follow the
recommended order. This improves readability and ensures a consistent
style guide for all developers working with Tailwind.